Deduplicate generated UI html utilities
